pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chestra"
version = "0.1.0"
description = "A modular, plugin-based workflow orchestrator for Pythonistas."
authors = [
    { name="Your Name", email="your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "PyYAML>=6.0",
    "requests>=2.28"
]
license = {text = "MIT"}

[project.optional-dependencies]
dev = ["ruff"] 